requirements.txt
pychromecast==13.1.0
pymediainfo
pysmb